IMS-28 Cleanup and testing

+++ src/main/java/com/qa/ims/controller/OrderController.java
@@ -1,10 +1,12 @@
 package com.qa.ims.controller;
 
+import java.text.DecimalFormat;
 import java.util.List;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import com.qa.ims.persistence.dao.ItemDAO;
 import com.qa.ims.persistence.dao.OrderDAO;
 import com.qa.ims.persistence.dao.OrderDetailDAO;
 import com.qa.ims.persistence.domain.Order;
@@ -18,15 +20,18 @@ import com.qa.ims.utils.Utils;
 public class OrderController implements CrudController<Order> {
 
 	public static final Logger LOGGER = LogManager.getLogger();
+	private DecimalFormat decim = new DecimalFormat("0.00");
 
 	private OrderDAO orderDAO;
     private OrderDetailDAO orderDetailDAO;
+	private ItemDAO itemDAO;
 	private Utils utils;
 
-	public OrderController(OrderDAO orderDAO, OrderDetailDAO orderDetailDAO, Utils utils) {
+	public OrderController(OrderDAO orderDAO, OrderDetailDAO orderDetailDAO, ItemDAO itemDAO, Utils utils) {
 		super();
 		this.orderDAO = orderDAO;
         this.orderDetailDAO = orderDetailDAO;
+		this.itemDAO = itemDAO;
 		this.utils = utils;
 	}
 
@@ -36,10 +41,18 @@ public class OrderController implements CrudController<Order> {
 	@Override
 	public List<Order> readAll() {
 		List<Order> orders = orderDAO.readAll();
-		for (Order order : orders) {
-			LOGGER.info(order);
-			for (OrderDetail orderDetail : orderDetailDAO.readOrder(order.getId())) {
-				LOGGER.info(orderDetail);
+		if (orders.isEmpty()) {
+			LOGGER.info("There are no orders in the database.");
+		} else {
+			Double totalCost;
+			for (Order order : orders) {
+				totalCost = 0d;
+				LOGGER.info(order.formattedString());
+				for (OrderDetail orderDetail : orderDetailDAO.readOrder(order.getId())) {
+					LOGGER.info(orderDetail.formattedString());
+					totalCost += orderDetail.getPrice();
+				}
+				LOGGER.info(String.format("Total Cost for Order %s: %s%n", order.getId(), decim.format(totalCost)));
 			}
 		}
 		return orders;
@@ -50,19 +63,38 @@ public class OrderController implements CrudController<Order> {
 	 */
 	@Override
 	public Order create() {
-		LOGGER.info("Please enter a customer ID");
-		Long custID = utils.getLong();
-        Order order = orderDAO.create(new Order(custID));
+		Long custID;
+		Order order = null;
+		while (order == null) {
+			LOGGER.info("Please enter a customer ID:");
+			custID = utils.getLong();
+        	order = orderDAO.create(new Order(custID));
+			if (order == null) {
+				LOGGER.info("Customer doesn't exist, please create one before adding it to an order\n");
+			}
+		}
 		LOGGER.info("Please enter a list of items IDs to add to the order, with their quantity\nEnter ! to stop");
         Long itemID = utils.getLong();
-        while (!itemID.equals((long) -1)) {
-			LOGGER.info("Please enter quantity");
-            Double quantity = utils.getDouble();
-			if (orderDetailDAO.readOrderItem(order.getId(), itemID) == null) {
-				orderDetailDAO.create(new OrderDetail(order.getId(), itemID, quantity));
-				LOGGER.info("Item added, enter next ID");
+		Long quantity;
+        while (!itemID.equals(-1l)) {
+			if (itemDAO.read(itemID) == null) {
+				LOGGER.info("That item doesn't exist, please create it before adding it to an order\nAdd another ID:");
 			} else {
-				LOGGER.info("This item already exists in this order, use update to change it\nEnter new ID:");
+				LOGGER.info("Please enter the quantity:");
+            	quantity = utils.getLong();
+				if (quantity.equals(-1l)) { 
+					break; 
+				}
+				if (quantity.equals(0l)) {
+					LOGGER.info("Quantity can't be 0\n");
+					continue;
+				}
+				if (orderDetailDAO.readOrderItem(order.getId(), itemID) == null) {
+					orderDetailDAO.create(new OrderDetail(order.getId(), itemID, quantity));
+					LOGGER.info("Item added, enter next ID:");
+				} else {
+					LOGGER.info("This item already exists in this order, use update to change it\nEnter new ID:");
+				}
 			}
             itemID = utils.getLong();
         }
@@ -74,21 +106,37 @@ public class OrderController implements CrudController<Order> {
 	 */
 	@Override
 	public Order update() {
-		LOGGER.info("Please enter the id of the order you would like to update");
+		LOGGER.info("Please enter the id of the order you would like to update:");
 		Long id = utils.getLong();
-		LOGGER.info("Please enter a new customer ID");
-		Long custID = utils.getLong();
-		Order order = orderDAO.update(new Order(id, custID));
+		Long custID;
+		Order order = null;
+		while (order == null) {
+			LOGGER.info("Please enter a new customer ID:");
+			custID = utils.getLong();
+        	order = orderDAO.update(new Order(id, custID));
+			if (order == null) {
+				LOGGER.info("Customer doesn't exist, please create one before adding it to an order\n");
+			}
+		}
 		LOGGER.info("Please enter a list of items IDs to add to the order, with their quantity\nIf the item is already in the order, it will be deleted\nEnter ! to stop");
         Long itemID = utils.getLong();
-        LOGGER.info("Please enter quantity");
-        Double quantity = utils.getDouble();
-		while (!itemID.equals((long) -1)) {
-            orderDetailDAO.update(new OrderDetail(order.getId(), itemID, quantity));
-		    LOGGER.info("Item added, enter next ID");
+		while (!itemID.equals(-1l)) {
+			if (itemDAO.read(itemID) == null) {
+				LOGGER.info("That item doesn't exist, please create it before adding it to an order\nAdd another ID:");
+			} else {
+				LOGGER.info("Please enter the quantity:");
+				Long quantity = utils.getLong();	
+				if (quantity.equals(-1l)) { 
+					break; 
+				}
+				if (quantity.equals(0l)) {
+					LOGGER.info("Quantity can't be 0\n");
+					continue;
+				}
+            	orderDetailDAO.update(new OrderDetail(order.getId(), itemID, quantity));
+		    	LOGGER.info("Item added, enter next ID:");
+			}
             itemID = utils.getLong();
-            LOGGER.info("Please enter quantity");
-            quantity = utils.getDouble();    
         }
 		return order;
 	}
@@ -100,17 +148,25 @@ public class OrderController implements CrudController<Order> {
 	 */
 	@Override
 	public int delete() {
-		LOGGER.info("Please enter the id of the order you would like to delete items from");
+		LOGGER.info("Please enter the id of the order you would like to delete items from:");
 		Long orderID = utils.getLong();
+		if (orderDetailDAO.readOrder(orderID).isEmpty()) {
+			LOGGER.info("This order doesn't contain any items, it will be deleted.\nTo add items to this order, re-create it.");
+			return orderDAO.delete(orderID);
+		}
 		LOGGER.info("Please enter the item ids you would like to delete from the order. Deleting all items will delete the order. Entering * will delete all items\nEnter ! to stop");
 		Long itemID = utils.getLong();
 		while (!itemID.equals(-1l)) {
 			if (itemID.equals(-2l)) {
-				LOGGER.info("Deleting all items");
-				for (OrderDetail orderDetail : orderDetailDAO.readOrder(orderID)) {
-					orderDetailDAO.delete(orderDetail.getID());
+				LOGGER.info("Are you sure you want to delete all items in order %d? This will delete all items in this order?\nY for Yes, N for No");
+				String input = utils.getString();
+				if (input.strip().equalsIgnoreCase("Y")) {		
+					LOGGER.info("Deleting all items:");
+					for (OrderDetail orderDetail : orderDetailDAO.readOrder(orderID)) {
+						orderDetailDAO.delete(orderDetail.getID());
+					}
+					return orderDAO.delete(orderID);
 				}
-				return orderDAO.delete(orderID);
 			}
 			orderDetailDAO.delete(orderDetailDAO.readOrderItem(orderID, itemID).getID());
 			LOGGER.info("Item deleted");
